Revert "Add back BillDocument – need to preserve full_text for search indexes"
Sorry for my inadequate commit message for 85e518a.

What prompted me to remove this was problem: datamade/la-metro-councilmatic#514 (comment)

We don't have any code in our signal handler to create or update CouncilmaticBillDocuments when the OCD Bill Documents are created. As such, when we went to go query for bills with documents we were getting nothing (because they were checking councilmatic bill document, not the ocd one).

I missed that this broke the tests.

This reverts commit 6ddf2a0a4a2bcf2f753b82b693824d466e16e6c0.